Added workaround for mission "Nines and AK's"
Changed the line "Requested file does not exist = 1" to "#Requested file does not exist = 1".
That stops the limit adjuster from showing an error after shooting the gas tank of the car.
Note: It's only a workaround to proceed flawless trough the mission and does not fix the cause of the error!
Normally you would need to close the errors and return to the game to continue the mission. 
For modders and testers: You should remove the "#" from the line to have a proper error report.